Avance 8
Se agrego un checkbox de auto completado de los campos de request postulation del usuario premium.
Se agrego validacion de si el usuario tiene postulaciones pendientes no podrá eliminar sus datos de postulación.
Cambio en el controlador de postulation donde al momento de postular y se guarde en la tabla Postulation se guarda en la tabla Postulation_status el status y reasons como null para manejar facilmente la validacion del eliminar datos si tiene postulaciones con estatus pendiente.
Se cambio el metodo de acceptPostulation y rejectPostulation para que actualice el status y reasons en el controlador de Vacancy.
Se cambio el metodo de cancelPostulation para que vuelva dejar en null el status y reasons en el controlador de Vacancy.
Si la postulacion tiene status distinto a sin postular el checkbox estara disabled.

+++ app/Http/Controllers/Portal/PostulationController.php
@@ -4,7 +4,11 @@ namespace App\Http\Controllers\Portal;
 
 use App\Http\Controllers\Controller;
 use App\Models\Postulation;
+use App\Models\PostulationStatus;
+use App\Models\PostulationUserData;
 use App\Models\Vacancy;
+use Carbon\Carbon;
+use GuzzleHttp\Client;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\Storage;
@@ -48,9 +52,16 @@ class PostulationController extends Controller implements \Illuminate\Routing\Co
         // Verificar si el usuario tiene los roles 'Super admin' o 'Admin' o si es "Postulante" con plan_id 2
         $hasUnlimitedApplications = $user->hasRole('Super admin') || $user->hasRole('Admin') || $isPostulantWithPlan2;
 
+        // Obtener la fecha y hora actual en Santiago
+        $client = new Client();
+        $response = $client->get('http://worldtimeapi.org/api/timezone/America/Santiago');
+        $data = json_decode($response->getBody()->getContents(), true);
+        $currentDateTime = Carbon::parse($data['datetime']);
+
         // Obtener el número de postulaciones del usuario actual en el mes actual si no tiene postulaciones ilimitadas
         $currentMonthApplications = $hasUnlimitedApplications ? 0 : Postulation::where('user_id', $user->id)
-            ->whereMonth('created_at', date('m'))
+            ->whereYear('created_at', $currentDateTime->year)
+            ->whereMonth('created_at', $currentDateTime->month)
             ->count();
 
         return view('portal.postulations.index', compact('vacancies', 'postulations', 'currentMonthApplications', 'hasUnlimitedApplications', 'isPostulantWithPlan2'));
@@ -61,7 +72,10 @@ class PostulationController extends Controller implements \Illuminate\Routing\Co
      */
     public function requestVacancy(Vacancy $vacancy)
     {
-        return view('portal.postulations.request_vacancy', compact('vacancy'));
+        $user = auth()->user();
+        $postulationDataAvailable = PostulationUserData::where('user_id', $user->id)->exists();
+        $postulations = Postulation::where('user_id', $user->id)->with('status')->get()->keyBy('vacancy_id');
+        return view('portal.postulations.request_vacancy', compact('vacancy', 'postulationDataAvailable', 'postulations'));
     }
 
     /**
@@ -69,13 +83,12 @@ class PostulationController extends Controller implements \Illuminate\Routing\Co
      */
     public function sendRequestVacancy(Request $request, Vacancy $vacancy)
     {
-        // Validar los datos del formulario
         $request->validate([
             'names' => 'required|string|max:255',
             'last_names' => 'required|string|max:255',
             'email' => 'required|email|max:255',
             'contact_number' => ['required', 'regex:/^\+569[0-9]{8}$/', 'phone:CL,mobile'],
-            'curriculum_vitae' => 'required|file|mimes:pdf,doc,docx|max:2048',
+            'curriculum_vitae' => 'required_if:autofill,0|file|mimes:pdf,doc,docx|max:2048',
             'fortalezas' => [
                 'required',
                 'string',
@@ -83,41 +96,55 @@ class PostulationController extends Controller implements \Illuminate\Routing\Co
             ],
             'reasons' => 'required|string',
         ], [
+            'curriculum_vitae.required_if' => 'El campo curriculum vitae es obligatorio.',
             'contact_number.regex' => 'El número de contacto debe comenzar con +569 seguido de ocho dígitos.',
             'contact_number.phone' => 'El número de contacto no es un número de celular chileno válido.',
             'fortalezas.regex' => 'Debe ingresar máximo diez fortalezas, separadas por comas.',
         ]);
 
-        // Manejo del archivo del currículum
         if ($request->hasFile('curriculum_vitae')) {
             $file = $request->file('curriculum_vitae');
             $originalName = $file->getClientOriginalName();
             $fileName = pathinfo($originalName, PATHINFO_FILENAME);
             $extension = $file->getClientOriginalExtension();
 
-            // Verificar si el archivo ya existe y añadir un número si es necesario
-            $filePath = "curriculums/{$originalName}";
+            $user = auth()->user();
+            $directory = $user->plan_id == 2 ? 'premium_user_curriculums' : 'curriculums';
+            $filePath = "{$directory}/{$originalName}";
             $counter = 1;
+
             while (Storage::disk('public')->exists($filePath)) {
-                $filePath = "curriculums/{$fileName}_{$counter}.{$extension}";
+                $filePath = "{$directory}/{$fileName}_{$counter}.{$extension}";
                 $counter++;
             }
 
-            // Guardar el archivo en la ubicación final
-            $file->storeAs('curriculums', basename($filePath), 'public');
+            $storedFilePath = $file->storeAs($directory, basename($filePath), 'public');
+        } else {
+            $userData = PostulationUserData::where('user_id', auth()->user()->id)->first();
+            $storedFilePath = $userData->curriculum_vitae ?? null;
         }
 
-        // Guardar los datos en la base de datos
-        Postulation::create([
+        if (!$storedFilePath) {
+            return back()->withErrors(['curriculum_vitae' => 'Debe proporcionar un currículum vitae.']);
+        }
+
+        $postulation = Postulation::create([
             'names' => $request->input('names'),
             'last_names' => $request->input('last_names'),
             'email' => $request->input('email'),
             'contact_number' => $request->input('contact_number'),
-            'curriculum_vitae' => $filePath ?? null, // Ruta del archivo guardado
+            'curriculum_vitae' => $storedFilePath,
             'strengths' => $request->input('fortalezas'),
             'reasons' => $request->input('reasons'),
             'vacancy_id' => $vacancy->id,
-            'user_id' => auth()->user()->id, // ID del usuario autenticado
+            'user_id' => auth()->user()->id,
+        ]);
+
+        // Agregar el registro en postulation_status con status y reasons como null
+        PostulationStatus::create([
+            'postulation_id' => $postulation->id,
+            'status' => null,
+            'reasons' => null,
         ]);
 
         // Mensaje de éxito
@@ -130,6 +157,62 @@ class PostulationController extends Controller implements \Illuminate\Routing\Co
         return redirect()->route('portal.postulations.index');
     }
 
+    public function getUserData(Request $request)
+    {
+        if (auth()->user()->plan_id != 2) {
+            return response()->json([], 403);
+        }
+
+        $userData = PostulationUserData::where('user_id', auth()->id())->first();
+        if ($userData) {
+            return response()->json($userData);
+        } else {
+            return response()->json([], 404);
+        }
+    }
+
+    public function saveSelected(Request $request)
+    {
+        $user = auth()->user();
+        $selectedVacancies = $request->input('selected_vacancies', []);
+
+        // Obtener los datos del usuario logueado de la tabla postulation_user_data
+        $userData = PostulationUserData::where('user_id', $user->id)->first();
+
+        if (!$userData) {
+            return response()->json([
+                'icon' => 'warning',
+                'title' => 'Datos de postulación faltantes',
+                'text' => 'Debe primero agregar sus datos de postulación para usar esta función.',
+                'redirect_url' => route('portal.premium_benefits.postulation_data'),
+            ]);
+        }
+
+        foreach ($selectedVacancies as $vacancyId) {
+            // Crear una nueva postulación
+            $postulation = Postulation::create([
+                'names' => $userData->names,
+                'last_names' => $userData->last_names,
+                'email' => $userData->email,
+                'contact_number' => $userData->contact_number,
+                'curriculum_vitae' => $userData->curriculum_vitae,
+                'strengths' => $userData->strengths,
+                'reasons' => $userData->reasons,
+                'vacancy_id' => $vacancyId,
+                'user_id' => $user->id,
+            ]);
+
+            // Crear el estado de la postulación con status y reasons como null
+            PostulationStatus::create([
+                'status' => null,
+                'reasons' => null,
+                'postulation_id' => $postulation->id,
+            ]);
+        }
+
+        return response()->json(['message' => 'Se han postulado correctamente a las vacantes seleccionadas.']);
+    }
+
     /**
      * Show the form for editing the specified resource.
      */
@@ -137,6 +220,9 @@ class PostulationController extends Controller implements \Illuminate\Routing\Co
     {
         // Cargar la relación vacancy para mostrar detalles de la vacante si es necesario
         $postulation->load('vacancy');
+
+        // Extraer solo el nombre del archivo de la ruta
+        $postulation->file_name = basename($postulation->curriculum_vitae);
 
         return view('portal.postulations.edit', compact('postulation'));
     }
@@ -165,6 +251,12 @@ class PostulationController extends Controller implements \Illuminate\Routing\Co
             'fortalezas.regex' => 'Debe ingresar máximo diez fortalezas, separadas por comas.',
         ]);
 
+        // Obtener el usuario logeado
+        $user = auth()->user();
+
+        // Determinar la carpeta de almacenamiento en función del plan_id
+        $folder = $user->plan_id == 2 ? 'premium_user_curriculums' : 'curriculums';
+
         // Manejo del archivo del currículum
         if ($request->hasFile('curriculum_vitae')) {
             $file = $request->file('curriculum_vitae');
@@ -173,15 +265,15 @@ class PostulationController extends Controller implements \Illuminate\Routing\Co
             $extension = $file->getClientOriginalExtension();
 
             // Verificar si el archivo ya existe y añadir un número si es necesario
-            $filePath = "curriculums/{$originalName}";
+            $filePath = "{$folder}/{$originalName}";
             $counter = 1;
             while (Storage::disk('public')->exists($filePath)) {
-                $filePath = "curriculums/{$fileName}_{$counter}.{$extension}";
+                $filePath = "{$folder}/{$fileName}_{$counter}.{$extension}";
                 $counter++;
             }
 
             // Guardar el archivo en la ubicación final
-            $file->storeAs('curriculums', basename($filePath), 'public');
+            $file->storeAs($folder, basename($filePath), 'public');
 
             // Borrar el archivo anterior si existe uno nuevo
             if ($postulation->curriculum_vitae) {
@@ -231,10 +323,31 @@ class PostulationController extends Controller implements \Illuminate\Routing\Co
      */
     public function destroy(Postulation $postulation)
     {
+        // Obtener el usuario que hizo la postulación
+        $user = $postulation->user;
+
         // Verificar si la aplicación tiene un archivo de currículum vitae
         if ($postulation->curriculum_vitae) {
-            // Eliminar el archivo del sistema de archivos
-            Storage::disk('public')->delete($postulation->curriculum_vitae);
+            // Verificar si el usuario tiene plan_id 1 y si el currículum vitae está en la carpeta 'curriculums'
+            if ($user->plan_id == 1 && strpos($postulation->curriculum_vitae, 'curriculums/') === 0) {
+                // Eliminar el archivo del sistema de archivos
+                Storage::disk('public')->delete($postulation->curriculum_vitae);
+            }
+
+            // Verificar si el usuario tiene plan_id 2
+            if ($user->plan_id == 2) {
+                // Obtener los datos de postulación del usuario
+                $postulationUserData = PostulationUserData::where('user_id', $user->id)->first();
+
+                // Verificar si el currículum de la postulación no coincide con el guardado en postulation_user_data
+                if ($postulationUserData && $postulation->curriculum_vitae !== $postulationUserData->curriculum_vitae) {
+                    // Verificar si el currículum vitae está en la carpeta 'premium_user_curriculums'
+                    if (strpos($postulation->curriculum_vitae, 'premium_user_curriculums/') === 0) {
+                        // Eliminar el archivo del sistema de archivos
+                        Storage::disk('public')->delete($postulation->curriculum_vitae);
+                    }
+                }
+            }
         }
 
         // Eliminar la aplicación de la base de datos
